+++ opennmt/models/sequence_classifier.py
@@ -3,6 +3,7 @@
 import tensorflow as tf
 
 from opennmt.models.model import Model
+from opennmt.utils.cell import last_encoding_from_state
 from opennmt.utils.misc import count_lines, print_bytes
 from opennmt.utils.losses import cross_entropy_loss
 
@@ -35,9 +36,10 @@ class SequenceClassifier(Model):
       ValueError: if :obj:`encoding` is invalid.
     """
     super(SequenceClassifier, self).__init__(
-        name, daisy_chain_variables=daisy_chain_variables, dtype=inputter.dtype)
+        name,
+        features_inputter=inputter,
+        daisy_chain_variables=daisy_chain_variables)
 
-    self.inputter = inputter
     self.encoder = encoder
     self.labels_vocabulary_file_key = labels_vocabulary_file_key
     self.encoding = encoding.lower()
@@ -46,23 +48,9 @@ class SequenceClassifier(Model):
       raise ValueError("Invalid encoding vector: {}".format(self.encoding))
 
   def _initialize(self, metadata):
-    self.inputter.initialize(metadata)
+    super(SequenceClassifier, self)._initialize(metadata)
     self.labels_vocabulary_file = metadata[self.labels_vocabulary_file_key]
     self.num_labels = count_lines(self.labels_vocabulary_file)
-
-  def _get_serving_input_receiver(self):
-    return self.inputter.get_serving_input_receiver()
-
-  def _get_features_length(self, features):
-    return self.inputter.get_length(features)
-
-  def _get_dataset_size(self, features_file):
-    return self.inputter.get_dataset_size(features_file)
-
-  def _get_features_builder(self, features_file):
-    dataset = self.inputter.make_dataset(features_file)
-    process_fn = self.inputter.process
-    return dataset, process_fn
 
   def _get_labels_builder(self, labels_file):
     labels_vocabulary = tf.contrib.lookup.index_table_from_file(
@@ -76,14 +64,14 @@ class SequenceClassifier(Model):
     }
     return dataset, process_fn
 
-  def _build(self, features, labels, params, mode, config):
+  def _build(self, features, labels, params, mode, config=None):
     with tf.variable_scope("encoder"):
-      inputs = self.inputter.transform_data(
+      inputs = self.features_inputter.transform_data(
           features,
           mode=mode,
-          log_dir=config.model_dir)
+          log_dir=config.model_dir if config is not None else None)
 
-      encoder_outputs, _, _ = self.encoder.encode(
+      encoder_outputs, encoder_state, _ = self.encoder.encode(
           inputs,
           sequence_length=self._get_features_length(features),
           mode=mode)
@@ -91,7 +79,7 @@ class SequenceClassifier(Model):
     if self.encoding == "average":
       encoding = tf.reduce_mean(encoder_outputs, axis=1)
     elif self.encoding == "last":
-      encoding = tf.squeeze(encoder_outputs[:, -1:, :], axis=1)
+      encoding = last_encoding_from_state(encoder_state)
 
     with tf.variable_scope("generator"):
       logits = tf.layers.dense(