+++ opennmt/tokenizers/tokenizer.py
@@ -2,13 +2,13 @@
 
 """Define base tokenizers."""
 
+import io
 import sys
 import os
 import abc
 import six
-import yaml
-
 import tensorflow as tf
+import yaml
 
 from opennmt.utils.misc import print_bytes
 
@@ -27,7 +27,7 @@ class Tokenizer(object):
     self._config = {}
     if configuration_file_or_key is not None and os.path.isfile(configuration_file_or_key):
       configuration_file = configuration_file_or_key
-      with open(configuration_file) as conf_file:
+      with io.open(configuration_file, encoding="utf-8") as conf_file:
         self._config = yaml.load(conf_file)
       self._configuration_file_key = None
     else:
@@ -48,7 +48,7 @@ class Tokenizer(object):
     """
     if self._configuration_file_key is not None:
       configuration_file = metadata[self._configuration_file_key]
-      with open(configuration_file) as conf_file:
+      with io.open(configuration_file, encoding="utf-8") as conf_file:
         self._config = yaml.load(conf_file)
 
   def tokenize_stream(self, input_stream=sys.stdin, output_stream=sys.stdout, delimiter=" "):