+++ opennmt/utils/losses.py
@@ -14,9 +14,16 @@ def _smooth_one_hot_labels(logits, labels, label_smoothing):
       dtype=logits.dtype)
 
 def _softmax_cross_entropy(logits, labels, label_smoothing, mode):
+  # Computes the softmax in full precision.
+  if logits.dtype.base_dtype != tf.float32:
+    logits = tf.cast(logits, tf.float32)
   if mode == tf.estimator.ModeKeys.TRAIN and label_smoothing > 0.0:
     smoothed_labels = _smooth_one_hot_labels(logits, labels, label_smoothing)
-    return tf.nn.softmax_cross_entropy_with_logits_v2(
+    if hasattr(tf.nn, "softmax_cross_entropy_with_logits_v2"):
+      cross_entropy_fn = tf.nn.softmax_cross_entropy_with_logits_v2
+    else:
+      cross_entropy_fn = tf.nn.softmax_cross_entropy_with_logits
+    return cross_entropy_fn(
         logits=logits, labels=smoothed_labels)
   else:
     return tf.nn.sparse_softmax_cross_entropy_with_logits(