+++ scripts/wmt/run_wmt_ende.sh
@@ -1,8 +1,6 @@
 #!/bin/bash
 
-export CUDA_VISIBLE_DEVICES=0,1,2,3
-
-PYTHONPATH=../.. python -m bin.main train_and_eval \
-  --model ../../config/models/transformer.py \
-  --config config/wmt_ende.yml \
-  --num_gpus 4
+onmt-main train_and_eval \
+          --model_type Transformer \
+          --config config/wmt_ende.yml \
+          --num_gpus 4